Revert "3rd Commit in the rebuild"
This reverts commit a3e947034b8d8d66bd4de266d529ebeba9136442.

+++ W-SmartShopSelution/SmartShopClassLibrary/DataModels/IncomeOrder/IncomeOrderProductModel.cs
@@ -15,12 +15,12 @@ namespace Library
 
         public decimal IncomePrice { get; set; }
 
-        public float Quantity { get; set; }
+        public int Quantity { get; set; }
 
         /// <summary>
         /// Return the total price of this IncomeOrderProduct
         /// </summary>
-        public decimal GetTotalProductPrice { get { return (decimal)Quantity * IncomePrice; }  }
+        public decimal GetTotalProductPrice { get { return Quantity * IncomePrice; }  }
 
     }
 }
